Made improvements to scoring and lives systems.
The score multiplier and lives systems now detect when you pass multiple thresholds in one go, and the multiplier now resets when a player projectile hits an object that is not an enemy.

Made a start on the leaderboards system, as well as the game over widget where the player enters their initials.

~George H